The project was reorganized to separate the api from the administration application, the entities were also separated in a new project to allow segmenting each project to a single responsibility.
The operation of the web api was completed, currently this web api gives us the facility to verify the availability of rooms, make a reservation, modify the reservation or cancel.

The availability table was modified to add the control fields related to the maximum days allowed in the reservation, the maximum number of days that can be reserved in advance and the minimum number of days to reserve.